Completed accounts page, except for the update command for the users405 table.

+++ account.php
@@ -59,23 +59,76 @@
 					$stmt->fetch();
 					$stmt->close();
 
-					echo "<div id='account_info_title'>Account Type:</div> <div id='account_info_field'>$role</div>";
-					echo "<div id='account_info_title'>First Name:</div> <div id='account_info_field'>$fname</div>";
-					echo "<div id='account_info_title'>Last Name:</div> <div id='account_info_field'>$lname</div>";
-					echo "<div id='account_info_title'>Email:</div> <div id='account_info_field'>$email</div>";
-					echo "<div id='account_info_title'>Street Address:</div> <div id='account_info_field'>$addr</div>";
-					echo "<div id='account_info_title'>Zip Code:</div> <div id='account_info_field'>$zip</div>";
-					echo "<div id='account_info_title'>City:</div> <div id='account_info_field'>$city</div>";
-					echo "<div id='account_info_title'>State:</div> <div id='account_info_field'>$state</div>";
+					if($_SERVER["REQUEST_METHOD"] == "POST") {
+						if(!empty($_POST["changeAcc"])) {
+							echo "<div id='account_info_title'>Account Type:</div> <div id='account_info_field'>
+								<input type='text' name='account_type'>
+								</div>";
+							echo "<div id='account_info_title'>First Name:</div> <div id='account_info_field'>
+								<input type='text' name='fname'>
+								</div>";
+							echo "<div id='account_info_title'>Last Name:</div>  <div id='account_info_field'>
+								<input type='text' name='lname'>
+								</div>";
+							echo "<div id='account_info_title'>Email:</div>  <div id='account_info_field'>
+								<input type='text' name='email'>
+								</div>";
+							echo "<div id='account_info_title'>Street Address:</div>  <div id='account_info_field'>
+								<input type='text' name='address'>
+								</div>";
+							echo "<div id='account_info_title'>Zip Code:</div>  <div id='account_info_field'>
+								<input type='text' name='zip_code'>
+								</div>";
+							echo "<div id='account_info_title'>City:</div>  <div id='account_info_field'>
+								<input type='text' name='city'>
+								</div>";
+							echo "<div id='account_info_title'>State:</div>  <div id='account_info_field'>
+								<input type='text' name='state'>
+								</div>";
+							#Update the DB for the changed fields
+							echo "<form method='post' action='./account.php'>";
+							echo "<input type='submit' value='Confirm Changes' name='confirmAcc' onClick='./account.php'>";
+							echo "</form>";
+						}
+						else if(!empty($_POST["confirmAcc"])){
+							echo "<span>Changes made to your account information.</span>";
+
+							echo "<div id='account_info_title'>Account Type:</div> <div id='account_info_field'>$role</div>";
+							echo "<div id='account_info_title'>First Name:</div> <div id='account_info_field'>$fname</div>";
+							echo "<div id='account_info_title'>Last Name:</div> <div id='account_info_field'>$lname</div>";
+							echo "<div id='account_info_title'>Email:</div> <div id='account_info_field'>$email</div>";
+							echo "<div id='account_info_title'>Street Address:</div> <div id='account_info_field'>$addr</div>";
+							echo "<div id='account_info_title'>Zip Code:</div> <div id='account_info_field'>$zip</div>";
+							echo "<div id='account_info_title'>City:</div> <div id='account_info_field'>$city</div>";
+							echo "<div id='account_info_title'>State:</div> <div id='account_info_field'>$state</div>";
+						}
+					}
+					else{
+						#$changes = 'htmlspecialchars($_SERVER["PHP_SELF"])';
+
+						echo "<div id='account_info_title'>Account Type:</div> <div id='account_info_field'>$role</div>";
+						echo "<div id='account_info_title'>First Name:</div> <div id='account_info_field'>$fname</div>";
+						echo "<div id='account_info_title'>Last Name:</div> <div id='account_info_field'>$lname</div>";
+						echo "<div id='account_info_title'>Email:</div> <div id='account_info_field'>$email</div>";
+						echo "<div id='account_info_title'>Street Address:</div> <div id='account_info_field'>$addr</div>";
+						echo "<div id='account_info_title'>Zip Code:</div> <div id='account_info_field'>$zip</div>";
+						echo "<div id='account_info_title'>City:</div> <div id='account_info_field'>$city</div>";
+						echo "<div id='account_info_title'>State:</div> <div id='account_info_field'>$state</div>";
+
+						echo "<form method='post' action='./account.php'>";
+						echo "<input type='submit' value='Change Account Information' name='changeAcc' onClick='./account.php'>";
+						echo "</form>";
+					}
 
 					
 					#implement ability to change.
 				?>
 				<br />
-				<form method="post" action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>">
-					<input type="submit" value="Change Account Information" onClick="./account.php">
+				
+					
 
 				</form>
+
 			</div>
 
 		</div>
